# apps/rag-contract/src/main/resources/openapi/rag-api.yaml
openapi: 3.0.3
info:
  title: RAG Systems API
  version: 1.0.0
  description: >
    Shared REST contract between the thin RAG TUI and the rag-* backend modules
    (rag-basic, rag-advanced, rag-agentic), plus the rag-memory and rag-webcrawler
    tools. Generated DTOs are consumed by every module so the contract stays in sync.

servers:
  - url: /
    description: Module-relative base path

tags:
  - name: documents
    description: Document ingestion into the active rag-module
  - name: query
    description: Retrieval search
  - name: chat
    description: Streaming chat
  - name: webcrawler
    description: Intelligent web fetching tool

paths:
  /api/documents/ingest:
    post:
      tags: [documents]
      operationId: ingestDocument
      summary: Ingest raw document content into the active rag-module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '201':
          description: Document ingested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Invalid request (blank content)

  /api/documents/ingest-url:
    post:
      tags: [documents]
      operationId: ingestUrl
      summary: >
        Orchestrated URL ingestion. The rag-module calls rag-webcrawler to fetch the
        page, then chunks (with its own chunker), embeds, stores, and returns when done.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestUrlRequest'
      responses:
        '201':
          description: Document ingested from URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Invalid URL

  /api/query:
    post:
      tags: [query]
      operationId: query
      summary: Similarity search over ingested documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Retrieved chunks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'

  /api/fetch:
    post:
      tags: [webcrawler]
      operationId: fetchPage
      summary: Fetch a single web page and extract its text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchRequest'
      responses:
        '200':
          description: Fetched page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          description: Invalid or unfetchable URL

  /api/fetch/links:
    post:
      tags: [webcrawler]
      operationId: fetchLinks
      summary: Fetch and rank candidate links (optional LLM-driven link selection)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchLinksRequest'
      responses:
        '200':
          description: Ranked candidate pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Page'

  /api/conversations:
    get:
      tags: [memory]
      operationId: listConversations
      summary: List stored conversations
      responses:
        '200':
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
    post:
      tags: [memory]
      operationId: createConversation
      summary: Create a new conversation
      responses:
        '201':
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'

  /api/conversations/{conversationId}/messages:
    get:
      tags: [memory]
      operationId: listMessages
      summary: List messages of a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
    post:
      tags: [memory]
      operationId: addMessage
      summary: Add a message to a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessage'
      responses:
        '201':
          description: Message added

components:
  schemas:
    IngestRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: Raw document text/body to ingest
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary document metadata (source, title, etc.)

    IngestResponse:
      type: object
      required: [documentId, chunkCount]
      properties:
        documentId:
          type: string
        chunkCount:
          type: integer
          format: int32

    IngestUrlRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        question:
          type: string
          description: Optional context used by rag-webcrawler for intelligent link selection

    QueryRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
        topK:
          type: integer
          format: int32
          default: 5
        documentId:
          type: string
          description: Optional scope to a single ingested document

    QueryResponse:
      type: object
      required: [question]
      properties:
        question:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/ChunkResult'

    ChunkResult:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        content:
          type: string
        index:
          type: integer
          format: int32

    FetchRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri

    FetchLinksRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        question:
          type: string

    Page:
      type: object
      required: [url, title, text]
      properties:
        url:
          type: string
        title:
          type: string
        text:
          type: string
        links:
          type: array
          items:
            type: string
          description: Candidate outbound links found on the page

    Conversation:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time

    ChatMessage:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time